lint: clear the clang-tidy backlog and pin the version in CI
Triaged all 55 findings clang-tidy 21 reports locally (39 of them on
clang-tidy 18, what CI installs -- the gap itself is the version-drift
problem noted below). Real bugs found and fixed:

- mktime/timegm: tm_hour*3600 and tm_min*60 were computed in `int`
  before widening to `long long`, in code that -- like tm_mon just
  above it -- is documented to accept out-of-range fields for
  normalization. A large caller-supplied tm_hour/tm_min overflows
  `int` (UB) before ever reaching the wider `secs` accumulator.
- asctime: same shape, tm_year + 1900 computed in `int` (`long` does
  not help here -- it's 32-bit on this LLP64 target, same as `int`).

Both widen the multiplicand/addend to `long long` before the op.

Everything else was triaged individually rather than blanket-silenced:
false positives from a deliberate `(x-x)/(x-x)` NaN idiom (pow.c,
trig.c) and a macro-expansion bit overlap (open.c) are NOLINT'd inline
with the reason; findings on defensively-clamped or taint-checker-
blind code (malloc.c, find_program.c, strtod.c, link.c) likewise.
pipe.c's char-to-WCHAR widening got an explicit (unsigned char) cast
since it was free and correct either way.

Checks disabled in .clang-tidy, with rationale added to the existing
per-check comment block: bugprone-switch-missing-default-case (several
switches here deliberately fall through), bugprone-suspicious-string-
compare (fires on ordinary `if (strcmp(...))`), bugprone-unsafe-
functions (flags *implementing* asctime_r), and cert-fio38-c /
misc-non-copyable-objects (flags the standard FILE storage and the
`malloc(sizeof *f)` idiom -- FILE isn't opaque here, ntlibc defines it).

Version drift: clang-tidy 21 (nix) finds 55, clang-tidy 18 (what CI's
apt install resolves to on ubuntu-24.04) finds 39 -- a gate whose
result depends on the toolchain image is fragile. Pinned rather than
loosened: tools/lint.sh now honors a CLANG_TIDY env var (default
`clang-tidy`), and ci.yml installs clang-tidy-18 explicitly and sets
CLANG_TIDY=clang-tidy-18, so an Ubuntu image bump can't silently
change what the gate checks.

Verified 0 findings with both clang-tidy 18 and 21 locally; `make
lint` exits 0 with all four stages run. `make check` 15/15 on x86_64
and i386, `make asan` 10/10, `make generated` drift-free.

+++ src/time/asctime.c
@@ -22,7 +22,12 @@ char *asctime_r(const struct tm *tm, char *buf)
 	n = __num_digits(p, 2, (unsigned)tm->tm_hour, 2, '0'); p += n; *p++ = ':';
 	n = __num_digits(p, 2, (unsigned)tm->tm_min, 2, '0'); p += n; *p++ = ':';
 	n = __num_digits(p, 2, (unsigned)tm->tm_sec, 2, '0'); p += n; *p++ = ' ';
-	n = __num_digits(p, 8, (unsigned long)(tm->tm_year + 1900), 4, '0'); p += n;
+	/* tm_year is `int` and unbounded (callers may hand back an out-of-range
+	 * value); widen to `long long` before the +1900 add so a caller-
+	 * supplied extreme value overflows there rather than in `int` --
+	 * plain `long` would not help here, since it is 32-bit on this
+	 * (LLP64) target same as `int`. */
+	n = __num_digits(p, 8, (unsigned long)((long long)tm->tm_year + 1900), 4, '0'); p += n;
 	*p++ = '\n';
 	*p = 0;
 	return buf;
